06-14-2022

+++ app/Http/Controllers/RoleController.php
@@ -11,11 +11,15 @@ class RoleController extends Controller
 {
     public function index()
     {
+        $this->authorize('view','users');
+
         return RoleResource::collection(Role::all());
     }
 
     public function store(Request $request)
     {
+        $this->authorize('edit','users');
+
         $role = Role::create($request->only('name'));
         $role->permissions()->attach($request->input('permissions'));
 
@@ -24,11 +28,15 @@ class RoleController extends Controller
 
     public function show($id)
     {
+        $this->authorize('view','users');
+
         return new RoleResource(Role::with('permissions')->find($id));
     }
 
     public function update(Request $request, $id)
     {
+        $this->authorize('edit','users');
+
         $role = Role::find($id);
         $role->update($request->only('name'));
         $role->permissions()->sync($request->input('permissions'));
@@ -38,6 +46,8 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('edit','users');
+
         Role::destroy($id);
 
         return \response(null, Response::HTTP_NO_CONTENT);
